Move classes into separate files and alter composer.json
The classes were moved into separate files.
I renamed all classes to be 'Fitbit'. Some classes were
FitBit with a capital B and others lowercase.
I edited composer.json because this doesn't directly
rely on the djchen project but it does directly rely
on the phpleague oauth2 library.
