+++ .gitignore
@@ -1,0 +1,6 @@
+/build
+/vendor
+composer.phar
+composer.lock
+.DS_Store
+.idea
