+++ src/FitbitException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace brulath\fitbit;
+
+
+class FitbitException extends \Exception
+{
+    protected $message = "Unknown Fitbit Exception";
+}
