+++ src/FitbitTokenExpiredException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace brulath\fitbit;
+
+
+class FitbitTokenExpiredException extends FitbitException {
+    protected $message = "Fitbit oauth token expired";
+}
