+++ src/FitbitTokenMissingException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace brulath\fitbit;
+
+
+class FitbitTokenMissingException extends FitbitException {
+    protected $message = "Fitbit oauth token missing";
+}
